folder structure finished
