+++ client/src/Routes.js
@@ -5,42 +5,40 @@ import {
   Switch,
   Redirect,
 } from "react-router-dom";
-// import Home from "./home";
-import Login from "./login";
-import Register from "./register";
-// import MyPwChange from "./pages/myPage/MyPwChange";
-// import MyProduct from "./pages/myPage/MyProduct";
-// import MyInfo from "./pages/myPage/MyInfo";
-// import SearchPage from "./pages/SearchPage";
-// import ArticlePage from "./pages/ArticlePage";
-// import OtherUserPage from "./pages/OtherUserPage";
-// import Changelayout from "./pages/myPage/mychage/ChangeLayout";
-// import UploadPage from "./pages/UploadPage";
-// import WishlistPage from "./pages/WishlistPage";
+import HomePage from "./home";
+import LoginPage from "./login";
+import RegisterPage from "./register";
+import MyPwChange from "./myPage/MyPwChange";
+import MyProduct from "./myPage/MyProduct";
+import MyInfo from "./myPage/MyInfo";
+import SearchPage from "./search";
+import ProductDetailPage from "./productDetail";
+import OtherUserPage from "./otherUser";
+import Changelayout from "./myPage/myChange/Changelayout";
+import UploadPage from "./upload";
 
 // Style Wrapper
 import { ThemeProvider } from "styled-components";
 import Theme from "./common/themes/Theme";
 import Global from "./common/themes/GlobalStyles";
-import productDetail from "./productDetail";
 
 export default function Routes() {
   return (
     <ThemeProvider theme={Theme}>
       <Router>
         <Switch>
-          {/* <Route exact path="/" component={HomePage} />
+          <Route exact path="/" component={HomePage} />
           <Route exact path="/mypwchange" component={MyPwChange} />
           <Route exact path="/myproduct" component={MyProduct} />
-          <Route exact path="/myinfo" component={MyInfo} /> */}
-          <Route exact path="/login" component={Login} />
-          <Route exact path="/register" component={Register} />
-          {/* <Route path="/search" component={SearchPage} /> */}
-          <Route path="/product/:product_id" component={productDetail} />
-          {/* <Route path="/user" component={OtherUserPage} />
-          <Route path="/upload" component={UploadPage} />
+          <Route exact path="/myinfo" component={MyInfo} />
+          <Route exact path="/login" component={LoginPage} />
+          <Route exact path="/register" component={RegisterPage} />
+          <Route path="/search" component={SearchPage} />
           <Route path="/myinfochange" component={Changelayout} />
-          <Route path="/wishlist" component={WishlistPage} /> */}
+          <Route path="/products/:product_id" component={ProductDetailPage} />
+          <Route path="/user" component={OtherUserPage} />
+          <Route path="/upload" component={UploadPage} />
+
           <Redirect from="*" to="/" />
         </Switch>
         <Global />
