--- client/src/containers/LoginContainer.js
@@ -1,1 +1,0 @@
-
