--- client/src/containers/OtherUserContainer.js
@@ -1,8 +1,0 @@
-import React from "react";
-import OtherUser from "../components/OtherUser";
-
-const OtherUserContainer = () => {
-  return <OtherUser />;
-};
-
-export default OtherUserContainer;
