+++ client/src/home/components/About.js
