--- client/src/pages/ArticlePage.js
@@ -1,8 +1,0 @@
-import React from "react";
-import ArticleContainer from "../containers/ArticleContainer";
-
-const ArticlePage = () => {
-  return <ArticleContainer />;
-};
-
-export default ArticlePage;
