--- client/src/pages/LoginPage.js
@@ -1,13 +1,0 @@
-import React from "react";
-import LoginContainer from "../containers/LoginContainer";
-import Layout from "../components/common/Layout";
-
-const LoginPage = () => {
-  return (
-    <Layout>
-      <LoginContainer />
-    </Layout>
-  );
-};
-
-export default LoginPage;
