--- client/src/pages/OtherUserPage.js
@@ -1,13 +1,0 @@
-import React from "react";
-import Layout from "../components/common/Layout";
-import OtherUserContainer from "../containers/OtherUserContainer";
-
-const OtherUserPage = () => {
-  return (
-    <Layout>
-      <OtherUserContainer />
-    </Layout>
-  );
-};
-
-export default OtherUserPage;
