--- client/src/pages/RegisterPage.js
@@ -1,13 +1,0 @@
-import React from "react";
-import Layout from "../components/common/Layout";
-import RegisterContainer from "../containers/RegisterContainer";
-
-const RegisterPage = () => {
-  return (
-    <Layout>
-      <RegisterContainer />
-    </Layout>
-  );
-};
-
-export default RegisterPage;
