--- client/src/pages/SearchPage.js
@@ -1,13 +1,0 @@
-import React from "react";
-import Layout from "../components/common/Layout";
-import SearchContainer from "../containers/SearchContainer";
-
-const SearchPage = () => {
-  return (
-    <Layout>
-      <SearchContainer />
-    </Layout>
-  );
-};
-
-export default SearchPage;
