--- client/src/pages/UploadPage.js
@@ -1,13 +1,0 @@
-import React from "react";
-import Layout from "../components/common/Layout";
-import UploadContainer from "../containers/UploadContainer";
-
-const UploadPage = () => {
-  return (
-    <Layout>
-      <UploadContainer />
-    </Layout>
-  );
-};
-
-export default UploadPage;
