--- client/src/pages/WishlistPage.js
@@ -1,13 +1,0 @@
-import React from "react";
-import Layout from "../components/common/Layout";
-import WishlistContainer from "../containers/WishlistContainer";
-
-const WishlistPage = () => {
-  return (
-    <Layout>
-      <WishlistContainer />
-    </Layout>
-  );
-};
-
-export default WishlistPage;
