+++ client/src/register/components/style.js
