+++ client/src/search/index.js
@@ -1,0 +1,8 @@
+import React from "react";
+import Container from "./Container";
+
+const SearchPage = () => {
+  return <Container />;
+};
+
+export default SearchPage;
