+++ client/src/upload/index.js
@@ -2,7 +2,7 @@ import React from "react";
 import Layout from "../common/components/Layout";
 import Container from "./Container";
 
-const RegisterPage = () => {
+const UploadPage = () => {
   return (
     <Layout>
       <Container />
@@ -10,4 +10,4 @@ const RegisterPage = () => {
   );
 };
 
-export default RegisterPage;
+export default UploadPage;
